% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprova_canvis_osm.R
\name{canvis_html}
\alias{canvis_html}
\title{canvis_html}
\usage{
canvis_html(x)
}
\arguments{
\item{x}{un objecte de \link[=compareDF]{compare_df}}
}
\value{
la taula html amb les diferències o una cadena de text indicant que no hi ha diferències.
}
\description{
Mostra diferències de taules en html si n'hi ha. Funció pensada per usar en fitxers \code{.qmd} o \code{.Rmd}.
}
