% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprova_canvis_osm.R
\name{comprova_canvis_osm}
\alias{comprova_canvis_osm}
\title{Consulta etiquetes d'OSM}
\usage{
comprova_canvis_osm(x, centre = FALSE)
}
\arguments{
\item{x}{un \code{data.frame} amb les columnes \code{osm_type} i \code{osm_id}.}

\item{centre}{si és \code{TRUE}, afegeix les coordenades del centre de l'objecte a les columnes \code{osm_center_lon} i
\code{osm_center_lat}.}
}
\value{
Retorna \code{x} amb les etiquetes dels objectes com a columnes. Si les columnes ja existien, actualitza els
valors de les etiquetes i conserva l'ordre de les columnes originals afegint les noves al final.
}
\description{
Consulta etiquetes d'OSM
}
\examples{
\dontrun{
comarques_canviades<- comarques
comarques_canviades<- comarques_canviades$name[1]<- "Malnom"
canvis <- comprova_canvis_osm(comarques)
compareDF::view_html(canvis)
}
}
