% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_etiquetes_osm.R
\name{consulta_etiquetes_osm}
\alias{consulta_etiquetes_osm}
\title{Consulta etiquetes d'OSM}
\usage{
consulta_etiquetes_osm(x, etiquetes, centre = FALSE)
}
\arguments{
\item{x}{un \code{data.frame} amb les columnes \code{osm_type} i \code{osm_id}.}

\item{etiquetes}{noms de les claus de les etiquetes a consultar. Si no s'especifica, s'afegeixen totes les etiquetes
dels objectes.}

\item{centre}{si és \code{TRUE}, afegeix les coordenades del centre de l'objecte a les columnes \code{osm_center_lon} i
\code{osm_center_lat}.}
}
\value{
Retorna \code{x} amb les etiquetes dels objectes com a columnes. Si les columnes ja existien, actualitza els
valors de les etiquetes i conserva l'ordre de les columnes originals afegint les noves al final.
}
\description{
Consulta etiquetes d'OSM
}
\examples{
\dontrun{
territoris_etiquetes_completes <- consulta_etiquetes_osm(territoris)
}
}
