% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram.R
\name{envia_missatge_telegram}
\alias{envia_missatge_telegram}
\title{Envia un missatge a un xat de Telegram}
\usage{
envia_missatge_telegram(
  missatge = "Hi ha canvis a [monitorOSM](https://osm-catalan.github.io/monitorOSM).",
  parse_mode = "Markdown"
)
}
\arguments{
\item{missatge}{Text del missatge.}

\item{parse_mode}{El missatge s'interpreta com a \code{Markdown} o \code{HTML}.

On i amb quin bot s'envia el missatge es configura amb les variables d'entorn \code{R_TELEGRAM_BOT_MONITOROSM_BOT} i
\code{TME_OSMCAT_CHATID}. Per usar la funció a github, cal configurar les accions perquè carreguin aquestes variables
d'entorn a partir dels secrets del repositori de github.}
}
\description{
Aquest repositori està configurat per enviar missatges al grup de Telegram de la
\verb{Comunitat en català d'OpenStreetMap} amb el bot \code{monitorOSM_bot}.
}
