% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_history.R
\name{get_osm_history}
\alias{get_osm_history}
\title{Retorna la història d'un element d'OSM}
\usage{
get_osm_history(osm_id, osm_type)
}
\arguments{
\item{osm_id}{una id d'una relació de la qual volem obtenir la història}

\item{osm_type}{string que especifica el tipus d'element ("node", "way" o "relation")}
}
\value{
En el cas que l'objecte sigui una relació, retorna una llista amb dos \code{data.frames}, un amb els membres de
totes les versions de l'objecte i un amb tots els tags, el número de versió, la id del changeset, la data del
canvi, el nom de l'usuari que el va fer i el seu uid. En cas que sigui una via, retorna una llista amb dos
\code{data.frame}, un amb tots els nodes que formen la via i un amb tots els tags. En cas que sigui un node, retorna un
\code{data.frame} amb informació sobre tots els canvis i tags'
}
\description{
Obté totes les versions de tots els elements i tags d'una relació d'OSM
}
\details{
Pensat per a realitzar un informe dels últims canvis realitzats a un objecte
}
\examples{
\dontrun{
id_la_selva <- 368462 # totes les versions de la relació que conté la comarca de la Selva
canvis <- get_osm_history(id_la_selva, "relation")
}
}
