% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edita_osm.R
\name{modifica_etiquetes_osm}
\alias{modifica_etiquetes_osm}
\title{Modifica etiquetes a OSM}
\usage{
modifica_etiquetes_osm(x, claus, comentari, ...)
}
\arguments{
\item{x}{\code{data.frame} amb els valors de les etiquetes que s'actualitzaran i les columnes \code{osm_type} i \code{osm_id} o
(\code{type} i \code{id}) que identifiquen els objectes d'OSM.}

\item{claus}{Claus de les etiquetes que es modificaran. Han de correspondre a columnes d'\code{x}.}

\item{comentari}{Comentari descriptiu del conjunt de canvis.}

\item{...}{Etiquetes del conjunt de canvis com a paràmetres amb noms (\code{clau = "valor"}).}
}
\value{
Retorna l'identificador del conjunt de canvis.
}
\description{
Obre un conjunt de canvis, edita les etiquetes i tanca el conjunt de canvis.
}
\details{
Cal identificar-se amb un usuari d'OSM i els canvis es faran amb aquest usuari i sota la seva responsabilitat.

Les etiquetes més habituals pels conjunts de canvis inclouen \code{source} i \code{hashtags}. En el marc del projecte,
\code{monitorOSM}, pot ser bona idea afegir el paràmetre \code{hashtags = "monitorOSM"} per poder fer un seguiment de les
edicions. Vegeu \url{https://wiki.openstreetmap.org/wiki/Changeset} per més detalls.

OSM és un projecte col·laboratiu on participen milers de persones. Reviseu bé els canvis, consensueu-los amb la
comunitat, afegiu comentaris informatius del conjunt de canvis i feu un bon ús d'aquesta funció.
}
\note{
Modificar les claus \code{osm_type}, \code{osm_id}, \code{id} i \code{type} amb aquesta funció pot donar problemes per col·lisions de
noms de columnes d'\code{x} i no és recomanat.
}
\examples{
# Restaura els noms de les comarques dels Països Catalans segons `comarques`
\dontrun{
modifica_etiquetes_osm(
  comarques,
  claus = "name",
  comentari = "Restaura els noms de les comarques dels Països Catalans", hashtags = "monitorOSM"
)
}
}
