% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{versio_canvis_html}
\alias{versio_canvis_html}
\title{Mostra les versions que introdueixen canvis en html}
\usage{
versio_canvis_html(x)
}
\arguments{
\item{x}{Una llista de canvis retornada per \code{\link[=cerca_versio_canvis]{cerca_versio_canvis()}}.}
}
\value{
Mostra les taules html amb informació del conjunt de canvis que ha modificat etiquetes.
}
\description{
Funció pensada per usar en fitxers \code{.qmd} o \code{.Rmd}.
}
