% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprova_canvis_osm.R
\name{comprova_canvis_osm}
\alias{comprova_canvis_osm}
\title{Comprova canvis a OSM}
\usage{
comprova_canvis_osm(x, centre = FALSE)
}
\arguments{
\item{x}{un \code{data.frame} amb les columnes \code{osm_type}, \code{osm_id} i etiquetes que vulguem comprovar.}

\item{centre}{si és \code{TRUE}, afegeix les coordenades del centre de l'objecte a les columnes \code{osm_center_lon} i
\code{osm_center_lat}.}
}
\value{
Retorna una comparació amb \code{\link[compareDF:compare_df]{compareDF::compare_df()}} de les diferències d'etiquetes dels objectes d'OSM d'\code{x}
respecte a les etiquetes actuals a OSM.
}
\description{
Compara objectes d'OSM respecte a una taula de referència
}
\details{
Només es comparen les etiquetes presents com a columnes d'\code{x} i la resta s'ometen.
}
\examples{
\dontrun{
comarques_canviades <- comarques
comarques_canviades <- comarques_canviades$name[1] <- "Malnom"
canvis <- comprova_canvis_osm(comarques)
compareDF::view_html(canvis)
}
}
