% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram.R
\name{missatge_resum_informes}
\alias{missatge_resum_informes}
\title{Crea missatge de resum d'informes}
\usage{
missatge_resum_informes(canvis, pagina_informe, etiquetes)
}
\arguments{
\item{canvis}{llista amb objectes produïts per \code{\link[=comprova_canvis_osm]{comprova_canvis_osm()}}. Si els elements tenen nom, s'usarà com
a nom de la fila de la taula resum.}

\item{pagina_informe}{nom de la pàgina de l'informe per generar l'enllaç a la web.}

\item{etiquetes}{Si no és \code{NULL}, quan es fa el recompte de canvis, només es tenen en compte les etiquetes que
quadren amb aquesta expressió regular.}
}
\value{
Text en format Markdown amb un missatge resum dels informes.
}
\description{
Crea missatge de resum d'informes
}
