% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram.R
\name{missatge_resum_informes}
\alias{missatge_resum_informes}
\title{Crea missatge de resum d'informes}
\usage{
missatge_resum_informes(..., pagina_informe, etiquetes)
}
\arguments{
\item{...}{objectes produits per \code{\link[=comprova_canvis_osm]{comprova_canvis_osm()}}. Si els parametres tenen nom, s'usarà com a nom de la fila
de la taula resum. Sinó, s'usarà el nom de la variable.}

\item{pagina_informe}{nom de la pàgina de l'informe per generar l'enllaç a la web.}

\item{etiquetes}{Si no és \code{NULL}, quan es fa el recomte de canvis, només té en compte les etiquetes que quadren amb
aquesta exrpessió regular.}
}
\value{
Text en format Markdown amb un missatge resum dels informes.
}
\description{
Crea missatge de resum d'informes
}
