% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{versio_canvis_html}
\alias{versio_canvis_html}
\title{Mostra les versions que introdueixen canvis en html}
\usage{
versio_canvis_html(x, titol)
}
\arguments{
\item{x}{Un dels elements de la llista retornada per \code{\link[=cerca_versio_canvis]{cerca_versio_canvis()}}.}

\item{titol}{Títol de la taula. S'espera una url i se li afegiran les etiquetes \\if{html}{\out{<a href=titol\\>}}titol\if{html}{\out{</a>}}}
}
\value{
La taula html amb informació del conjunt de canvis que ha modificat etiquetes.
}
\description{
Mostra les versions que introdueixen canvis en html
}
